requirements.txt
# QeMLflow Dependencies - Updated for Modern Compatibility
# =======================================================

# Core ML and Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning Frameworks (Updated versions for compatibility)
tensorflow>=2.16.0,<2.17.0  # Avoid experimental warnings
torch>=2.0.0
torchvision>=0.15.0

# DISTRIBUTED ML TRAINING - Best-in-Class Libraries
# ==================================================
ray[default]>=2.5.0          # Primary distributed ML platform
ray[tune]>=2.5.0             # Hyperparameter optimization
ray[train]>=2.5.0            # Distributed training
horovod>=0.28.0              # Multi-GPU/multi-node training
dask[distributed]>=2023.5.0  # Alternative distributed computing
pytorch-lightning>=2.1.0     # High-level PyTorch training
tensorflow-probability>=0.22.0  # Probabilistic programming

# HYPERPARAMETER OPTIMIZATION - Advanced Algorithms
# ==================================================
optuna>=3.4.0                # State-of-the-art Bayesian optimization
hyperopt>=0.2.7              # Tree Parzen Estimator optimization
ax-platform>=0.3.4           # Facebook's Adaptive Experimentation
scikit-optimize>=0.9.0       # Gaussian process optimization
nni>=3.0                     # Microsoft Neural Network Intelligence
flaml>=2.1.0                 # Microsoft's fast AutoML

# PERFORMANCE MONITORING - Production-Grade Observability
# =======================================================
mlflow>=2.8.0                # Model lifecycle management
wandb>=0.15.0                # Real-time experiment tracking
tensorboard>=2.14.0          # TensorFlow visualization
prometheus-client>=0.18.0    # Metrics collection
grafana-api>=1.0.3          # Dashboard creation
psutil>=5.9.0                # System resource monitoring
gpustat>=1.1.0               # GPU monitoring
neptune-client>=1.7.0        # Alternative experiment tracking

# AUTO-ML AND MODEL SELECTION
# ============================
auto-sklearn>=0.15.0         # Automated machine learning
pycaret>=3.1.0               # Low-code ML platform
h2o>=3.42.0                  # Enterprise AutoML
tpot>=0.12.0                 # Genetic programming AutoML

# Cheminformatics and Drug Discovery
rdkit>=2023.3.1
deepchem>=2.8.0
biopython>=1.81

# QUANTUM COMPUTING - Multi-Backend Flexibility
# ==============================================
pennylane>=0.32.0            # Best quantum ML library
qiskit>=1.0.0                # IBM's quantum framework
qiskit-aer>=0.13.0          # High-performance simulators
cirq>=1.2.0                 # Google's quantum framework
qulacs>=0.6.0               # Ultra-fast quantum simulator
pytket>=1.20.0              # Cambridge Quantum Computing

# KNOWLEDGE INTEGRATION - Enterprise NLP
# ======================================
transformers>=4.35.0         # State-of-the-art language models
datasets>=2.14.0             # Hugging Face datasets
neo4j>=5.13.0                # Graph database for knowledge graphs
spacy>=3.7.0                 # Industrial-strength NLP
py2neo>=2021.2.3            # Python Neo4j client
networkx>=3.2.0             # Graph analysis
sentence-transformers>=2.2.2 # Semantic text embeddings
langchain>=0.0.340          # LLM application framework

# DeepChem Optional Dependencies (to reduce warnings)
dgl>=1.1.0  # Deep Graph Library
lightning>=2.0.0  # PyTorch Lightning (renamed from pytorch-lightning)
jax>=0.4.0  # JAX for advanced ML
jaxlib>=0.4.0  # JAX backend
torch-geometric>=2.3.0  # PyTorch Geometric

# UNCERTAINTY QUANTIFICATION - Advanced Methods
# =============================================
pyro-ppl>=1.8.6              # Probabilistic programming
gpytorch>=1.11.0             # Gaussian processes
botorch>=0.9.4               # Bayesian optimization
uncertainty-toolbox>=0.1.1   # Uncertainty metrics
conformal-prediction>=0.1.9  # Distribution-free uncertainty

# Molecular Dynamics and Simulation
openmm>=8.0.0
mdtraj>=1.9.8
pymol-open-source>=2.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
py3Dmol>=2.0.0
nglview>=3.0.3

# Molecular Descriptors and QSAR
mordred>=1.2.0
chembl-webresource-client>=0.10.8

# Data Processing and Analysis
jupyter>=1.0.0
ipywidgets>=8.0.0
tqdm>=4.64.0

# Classical Quantum Chemistry (for comparison)
pyscf>=2.3.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
joblib>=1.1.0
mypy>=1.6.0
bandit>=1.7.5                # Security scanning
vulture>=2.7                 # Dead code detection
flake8>=6.1.0                # Core linting
flake8-bugbear>=23.9.0       # Additional flake8 checks
flake8-comprehensions>=3.14.0 # List/dict comprehension checks
pre-commit>=3.5.0            # Git hooks
